Fast, parallel implementation of grid data convolution

A fast, parallel implementation of convolutions for grid-type data (matrices, rasters and other grid based objects).

pFocal(
  data,
  kernel,
  edge_value = 0,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...
)

Arguments

data
[matrix-type] Grid to compute onto.

kernel
[matrix] Computation kernel (neighborhood).

edge_value
[numeric] Numeric value, NA or NaN.The value to give at the edge of the kernel matrix when the moving window overflows the data grid.

transform_function
[character] The function to apply to the cell values covered by the kernel. For possible values, see pFocal_transform_info(). Default to "MULTIPLY".

reduce_function
[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pFocal_reduce_info(). Default to "SUM".

mean_divider
[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pFocal_mean_divisor_info(). Default to "ONE" (for no division).

variance
[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pFocal_variance_info()). Default to FALSE (just returns the value at each point).

na.rm
[NA OR character] The behavior to adopt for dealing with missing values, default to NA. Far possible values see pFocal_nan_policy_info().

mp
[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation
[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...
None used.

Value

The updated, convoluted grid.

Details

Note that the memory allocation for the output is of size sizeof(double) * ncol * nrow and for the intermediate values, sizeof(double) * (ncol + kernel_ncol) * (nrow + kernel_nrow/2).

Examples

data <- matrix(nrow = 10, ncol = 10, data = runif(10 * 10))
kernel <- matrix(1 / 9, nrow = 3, ncol = 3)
pFocal(data = data, kernel = kernel)

#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2182916 0.3258775 0.2527341 0.2671193 0.2458033 0.3393838 0.2992955
#>  [2,] 0.2888474 0.4781354 0.3716665 0.4535449 0.3699351 0.4934283 0.3851602
#>  [3,] 0.2357633 0.4146141 0.3668982 0.5081472 0.4427955 0.4446536 0.3407370
#>  [4,] 0.1691471 0.3815711 0.4538991 0.5251250 0.4460217 0.4555446 0.3984436
#>  [5,] 0.1721539 0.3852670 0.4871181 0.5541764 0.5415655 0.4944364 0.4226781
#>  [6,] 0.2192341 0.4162904 0.4880748 0.5580485 0.5251170 0.5382187 0.4444151
#>  [7,] 0.2129786 0.3599986 0.4525813 0.6595944 0.7011170 0.6560704 0.4817328
#>  [8,] 0.2640364 0.4059789 0.4409703 0.6493044 0.6165610 0.6387275 0.5346230
#>  [9,] 0.3581201 0.5708526 0.5209242 0.6234248 0.5365613 0.5387686 0.4778119
#> [10,] 0.3188503 0.4726703 0.3495409 0.3490358 0.2418184 0.2891475 0.2760574
#>            [,8]      [,9]     [,10]
#>  [1,] 0.3384207 0.3436525 0.2527399
#>  [2,] 0.4339181 0.4334291 0.3393986
#>  [3,] 0.3552445 0.3923701 0.2961968
#>  [4,] 0.3797337 0.4442803 0.3460492
#>  [5,] 0.4099683 0.4857687 0.3884261
#>  [6,] 0.4803792 0.4549440 0.3674920
#>  [7,] 0.4636233 0.4571118 0.3507842
#>  [8,] 0.5682750 0.4988254 0.2994957
#>  [9,] 0.4567823 0.4530558 0.2528551
#> [10,] 0.3106083 0.2744016 0.1364135