Fast methods for common kernel computations

Methods wrapping pFocal to implement common kernel computations with default argument values.

fast_gaussian_radius_pFocal(
  data,
  vertical_radius,
  vertical_sd = 1,
  horizontal_radius = vertical_radius,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

fast_gaussian_confidence_pFocal(
  data,
  vertical_r0 = 0.05,
  vertical_sd = 1,
  horizontal_r0 = vertical_r0,
  horizontal_sd = vertical_sd,
  tail_included = TRUE,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

fast_binomial_pFocal(
  data,
  vertical_radius,
  horizontal_radius = vertical_radius,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

fast_abs_rectangle_pFocal(
  data,
  height,
  width = height,
  value = 1,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

fast_seperated_pFocal(
  data,
  kernel_list,
  na.rm = NA,
  mp = TRUE,
  debug_use_r_implementation = FALSE,
  ...,
  transform_function = "MULTIPLY",
  reduce_function = "SUM",
  mean_divider = "ONE",
  variance = FALSE
)

Arguments

data
[matrix-type] Grid to compute onto.

vertical_radius
[numeric] The kernel's radius in the vertical dimension.

vertical_sd
[numeric] The kernel's standard deviation in the vertical dimension.

horizontal_radius
[numeric] The kernel's radius in the horizontal dimension.

horizontal_sd
[numeric] The kernel's standard deviation in the horizontal dimension.

tail_included
[logical] TODO

na.rm
[NA OR character] The behavior to adopt for dealing with missing values, default to NA. Far possible values see pFocal_nan_policy_info().

mp
[logical] Whether to use the open_mp implementation, default to TRUE.

debug_use_r_implementation
[logical] Used for debugging purposes whether to use the slow R implementation instead of the fass C++ code. Default to FALSE.

...
None used.

transform_function
[character] The function to apply to the cell values covered by the kernel. For possible values, see pFocal_transform_info(). Default to "MULTIPLY".

reduce_function
[character] The function to apply to the kernel values after the function passed in transform_function has been applied (the function that summarize the data). For possible values, see pFocal_reduce_info(). Default to "SUM".

mean_divider
[character] Optional, allows to specify how the final value at each cell is divided by a value that can be function of the intermediate data resulting of applying transform_function. For possible values, see pFocal_mean_divisor_info(). Default to "ONE" (for no division).

variance
[logical] Whether to return the "variance" of the intermediate values at each point (for more details please see pFocal_variance_info()). Default to FALSE (just returns the value at each point).

vertical_r0
[numeric] The kernel's r0 (exponential) in the vertical dimension.

horizontal_r0
[numeric] The kernel's r0 (exponential) in the horizontal dimension.

height
[numeric] For rectangular kernels, the height of the rectangle.

width
[numeric] For rectangular kernels, the width of the rectangle.

value
[numeric] For single value matrices, the value.

kernel_list
[list] A list of kernels computed from functions in kernel-gaussian, kernel-binomial, kernel-circular, kernel-distance, kernel-exponential.

Value

The updated, convoluted grid.